backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
